% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bringObjectTS.R
\name{bringObjectTS}
\alias{bringObjectTS}
\title{Bring objects that represent time series into a list}
\usage{
bringObjectTS(path, rastersNamePattern)
}
\arguments{
\item{path}{path to the directory containing the files to be brought as list.}

\item{rastersNamePattern}{Character vector of the name pattern to be used for the search.}
}
\value{
A list of the objects that were read from disk. Currently the
function only works with \code{.tif} (i.e. raster) and \code{.rds} objects.
}
\description{
Bring objects that represent time series into a list
}
\author{
Tati Micheletti
}
