% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbindFromList.R
\name{cbindFromList}
\alias{cbindFromList}
\title{Column binds a list of \code{data.frame}s or \code{data.table}s}
\usage{
cbindFromList(lst)
}
\arguments{
\item{lst}{List of \code{data.table}s or \code{data.frame}s to column bind.}
}
\value{
Returns a merged \code{data.table} with unique columns
}
\description{
Column binds a list of \code{data.frame}s or \code{data.table}s
}
\author{
Tati Micheletti
}
