% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDynamicLayersRSF.R,
%   R/createStaticLayersRSF.R
\name{createDynamicLayersRSF}
\alias{createDynamicLayersRSF}
\alias{createStaticLayersRSF}
\title{Create the necessary layers for predictive caribou RSF module.}
\usage{
createDynamicLayersRSF(
  ageMap,
  biomassMap,
  biomassMapName,
  oldBurnTime,
  oldBurnName,
  newBurnName,
  roadDensity,
  roadDensityName,
  waterRaster,
  waterRasterName,
  RTM
)

createStaticLayersRSF(
  elevation,
  vrug,
  LCC,
  shrubName,
  herbName,
  elevationName,
  vrugName,
  reclassLCC05,
  dynamicLayers,
  RTM,
  destinationPath
)
}
\arguments{
\item{ageMap}{\code{RasterLayer} of forest age.}

\item{biomassMap}{\code{RasterLayer}of forest biomass.}

\item{biomassMapName}{character. Name of the forest biomass layer.}

\item{oldBurnTime}{numeric. Definition of the initial interval considered to be old burn.
The end of this time is 20 years later (i.e. 40-60 years).}

\item{oldBurnName}{character. Name of the old burn layer in the model.}

\item{newBurnName}{character. Name of the new burn layer in the model.}

\item{roadDensity}{Anthropogenic disturbance (raster) layer. Currently, road density layer used
for both RSF and demographic models.}

\item{roadDensityName}{character. Name of the road Density layer in the model.}

\item{waterRaster}{\code{RasterLayer} indicating water bodies.}

\item{waterRasterName}{character. Name of the water layer in the model.}

\item{RTM}{\code{RasterLayer} template for these layers to match.}

\item{elevation}{\code{RasterLayer} of elevation}

\item{vrug}{\code{RasterLayer} of ruggedness}

\item{LCC}{\code{RasterLayer} of landcover classes 2005}

\item{shrubName}{character. Name of the shrub layer in the model.}

\item{herbName}{character. Name of the herb layer in the model.}

\item{elevationName}{character. Name of the elevation layer in the model.}

\item{vrugName}{character. Name of the vrug layer in the model.}

\item{reclassLCC05}{List with reclassification for LCC05 values
(i.e. LCC05 classes that should be classified as shrub or herbs)}

\item{dynamicLayers}{\code{RasterStack} to be stacked with the \code{staticLayers}
created in the current function to be passed for the model prediction.}

\item{destinationPath}{TODO}
}
\value{
\code{RasterStack} of layers

\code{RasterStack}
}
\description{
Intended for use within a SpaDES module.
}
\author{
Tati Micheletti
}
