% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEquation.R
\name{createEquation}
\alias{createEquation}
\title{Write an equation based on a \code{data.frame} (or \code{data.table}), bootstrapping a
set of covariates around the standard error present in the table.}
\usage{
createEquation(model, replicates = 100)
}
\arguments{
\item{model}{\code{data.frame}with columns \code{Coefficient}, and \code{Value} and \code{StdErr} for a given parameter.
The model table NEEDS to have at least two rows, one being the \code{Intercept}
as \code{Coefficient}. If \code{Intercept} is not found, it enters \code{browser()} mode.}

\item{replicates}{numeric. Number of repetitions to be added to the bootstrapping.}
}
\value{
As with \code{archivist::cache}, returns the value of the
function call or the cached version (i.e., the result from a previous call
to this same cached function with identical arguments).
}
\description{
Write an equation based on a \code{data.frame} (or \code{data.table}), bootstrapping a
set of covariates around the standard error present in the table.
}
\author{
Tati Micheletti
}
