% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModels.R
\name{createModels}
\alias{createModels}
\title{Create caribou models based on 2011 ECCC report on population demographics and RSF.}
\usage{
createModels(caribouCoefTable = NULL, modelsToUse = "M3")
}
\arguments{
\item{caribouCoefTable}{\code{data.table} of the coefficient values for all models available.
If not provided, will use as default the one from ECCC 2011 for
population demographics.}

\item{modelsToUse}{Character string of the model to be used. Currently only \code{"M3"} and \code{"M7"}
(demographics) and \code{"TaigaPlains"} (RSF) have been implemented. Default \code{"M3"}.}
}
\value{
named list of the quote of the model that can be parsed to generate the simulated data
bootstrapped using the model coefficients.
}
\description{
Uses \code{createEquation} internally.
}
\author{
Tati Micheletti
}
