% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShrubHerbLayers.R
\name{createShrubHerbLayers}
\alias{createShrubHerbLayers}
\title{Creates the necessary shrub or herb layers for predictive caribou RSF module.}
\usage{
createShrubHerbLayers(
  landCoverECCC,
  reclassLCC05,
  layerName,
  includeCrops = FALSE
)
}
\arguments{
\item{landCoverECCC}{\code{RasterLayer} LCC2005 reclassified to ECCC landcover.}

\item{reclassLCC05}{List with reclassification for LCC05 values
(i.e. LCC05 classes that should be classified as shrub or herbs)}

\item{layerName}{character. Name of the shrub layer in the model.}

\item{includeCrops}{Logical. Should this layer include crops?}
}
\value{
\code{RasterLayer} of herbs or shrubs
}
\description{
Creates the necessary shrub or herb layers for predictive caribou RSF module.
}
\author{
Tati Micheletti
}
