% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestAgePlot.R
\name{forestAgePlot}
\alias{forestAgePlot}
\title{Plots of forest age}
\usage{
forestAgePlot(
  dataPath,
  typeSim,
  addCaribousuitability = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{dataPath}{character. Path to data}

\item{typeSim}{character. Which simulation is it? i.e., \code{'LandR_SCFM'} or \code{'LandR.CS_fS'}}

\item{addCaribousuitability}{logical. Should the plot show which forest ages are better or worse
for Caribou in a colour-coded way?}

\item{overwrite}{logical.}
}
\value{
plot
}
\description{
Plots of forest age
}
\author{
Tati Micheletti
}
