% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayers.R
\name{getLayers}
\alias{getLayers}
\title{Get the necessary layers for predictive modules such as \code{caribouPopGrowth} and \code{caribouRSF}.}
\usage{
getLayers(
  currentTime,
  cohortData,
  pixelGroupMap,
  startTime,
  endTime,
  recoveryTime = 40,
  listSACaribou,
  anthropogenicLayer,
  roadDensity,
  waterRaster,
  isRSF = FALSE,
  decidousSp = NULL,
  oldBurnTime = NULL,
  elevation = NULL,
  vrug = NULL,
  LCC05 = NULL,
  reclassLCC05 = NULL,
  rasterToMatch = NULL,
  destinationPath
)
}
\arguments{
\item{currentTime}{numeric. Current time being used (i.e. \code{time(sim)}).}

\item{cohortData}{\code{data.table}. Output from \code{LandR_Biomass} module.}

\item{pixelGroupMap}{raster to identify the \code{cohortData}.}

\item{startTime}{numeric. start time of the simulation. Needed to verify and potentially adjust relative simulation times.}

\item{endTime}{numeric. end time of the simulation. Needed to verify and potentially adjust relative simulation times.}

\item{recoveryTime}{numeric. Recovery time in years that the forest needs to support Caribou. Default = 40.}

\item{listSACaribou}{list of shapefiles with polygons for which we want to calculate lambda for the caribou demographic model.}

\item{anthropogenicLayer}{Anthropogenic disturbance (raster) layer. Currently, 500 m buffered anthropogenic disturbance for demographic models.}

\item{roadDensity}{Anthropogenic disturbance (raster) layer. Currently, road density layer used for RSF models.}

\item{waterRaster}{\code{RasterLayer} indicating water bodies.}

\item{isRSF}{logical. Identify if it should get the layers for the RSF or demographic model.}

\item{decidousSp}{binary raster layer indicating if the dominant biomass in a pixel belongs to a deciduous species.}

\item{oldBurnTime}{numeric. Definition of the initial interval considered to be old burn. The end of this time is 20 years later (i.e. 40-60 years).}

\item{elevation}{\code{RasterLayer} of elevation}

\item{vrug}{\code{RasterLayer} of ruggedness}

\item{LCC05}{\code{RasterLayer} of landcover classes 2005}

\item{reclassLCC05}{List with reclassification for LCC05 values (i.e. LCC05 classes that should be classified as shrub or herbs)}

\item{rasterToMatch}{\code{RasterLayer} template for these layers to match.}

\item{destinationPath}{TODO}
}
\value{
TODO
}
\description{
Get the necessary layers for predictive modules such as \code{caribouPopGrowth} and \code{caribouRSF}.
}
\author{
Tati Micheletti
}
