% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReclassifyMatrix.R
\name{makeReclassifyMatrix}
\alias{makeReclassifyMatrix}
\title{Create a matrix to use for reclassification of rasters}
\usage{
makeReclassifyMatrix(table, originalCol, reclassifiedTo)
}
\arguments{
\item{table}{\code{data.frame} or \code{data.table} that has the original classification
and the desired classification to replace the first.}

\item{originalCol}{character. Column in the table that identifies the original classification.}

\item{reclassifiedTo}{character. Column in the table that identifies the desired classification.}
}
\value{
Matrix to use for reclassification of rasters
}
\description{
Create a matrix to use for reclassification of rasters
}
\author{
Tati Micheletti
}
