% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleSticker.R
\name{modulesAvailable}
\alias{modulesAvailable}
\title{Show the modules that have stickers available}
\usage{
modulesAvailable()
}
\value{
A character vector of all modules that have stickers available.
}
\description{
Show the modules that have stickers available
}
\author{
Tati Micheletti
}
