% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAndBringOn.R
\name{nameAndBringOn}
\alias{nameAndBringOn}
\title{Name a raster, \code{postProcess} it using RTM, and brings it to memory}
\usage{
nameAndBringOn(ras, name, RTM = NULL)
}
\arguments{
\item{ras}{\code{RasterLayer}.}

\item{name}{character. Name of the raster layer.}

\item{RTM}{\code{RasterLayer} template.
If a RTM is passed, the function masks the raster to it, converting non-NA to 0.}
}
\value{
\code{RasterStack}
}
\description{
Name a raster, \code{postProcess} it using RTM, and brings it to memory
}
\author{
Tati Micheletti
}
