% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBurnSummary.R
\name{plotBurnSummary}
\alias{plotBurnSummary}
\title{Plots summary of burns}
\usage{
plotBurnSummary(
  dataPath,
  typeSim,
  lastYear,
  theObject = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{dataPath}{character. Path to data}

\item{typeSim}{character. Which simulation is it? i.e., \code{'LandR_SCFM'} or \code{'LandR.CS_fS'}}

\item{lastYear}{TODO}

\item{theObject}{TODO}

\item{overwrite}{logical. Default FALSE}
}
\value{
plot
}
\description{
Plots summary of burns
}
\author{
Tati Micheletti
}
