% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLeadingVegetationType.R
\name{plotLeadingVegetationType}
\alias{plotLeadingVegetationType}
\title{Plots Leading Vegetation Type using \code{cohortData} and \code{pixelGroupMap}}
\usage{
plotLeadingVegetationType(
  dataPath,
  typeSim,
  colNA = "grey85",
  saveRAS = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{dataPath}{character. Path to data}

\item{typeSim}{character. Which simulation is it? i.e. \code{'LandR_SCFM'} or \code{'LandR.CS_fS'}.}

\item{colNA}{character. The colour to use for \code{NA}.}

\item{saveRAS}{logical. Save the raster for posterior use?}

\item{overwrite}{logical.}
}
\value{
plot
}
\description{
Plots Leading Vegetation Type using \code{cohortData} and \code{pixelGroupMap}
}
\author{
Tati Micheletti
}
