% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInputsLayers_DUCKS.R
\name{prepInputsLayers_DUCKS}
\alias{prepInputsLayers_DUCKS}
\title{Prepare DUCKS layer}
\usage{
prepInputsLayers_DUCKS(
  destinationPath,
  lccLayer = "2005",
  url = NULL,
  archive = NULL,
  targetFile = NULL,
  studyArea = NULL,
  rasterToMatch = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{destinationPath}{Path where to save the downloaded file.}

\item{lccLayer}{Which year should be used as a base for the vegetation layer? Default to 2005}

\item{url}{The url from where the layer should be downloaded from.
If \code{NULL}, the default is used.}

\item{archive}{TODO}

\item{targetFile}{TODO}

\item{studyArea}{Study area for which the layer should be cropped to}

\item{rasterToMatch}{TODO}

\item{overwrite}{Logical indicating whether to overwrite the previously processed file.}
}
\value{
RasterLayer
}
\description{
Intended to prepare the DUCKS Unlimited Hybrid Wetland v. 2.1 layer for different purposes.
The output is a \code{RasterLayer} cropped and reprojected to the \code{studyArea}, and resampled to the
\code{rasterToMatch} if any of these are provided.
}
