% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviseSpeciesTraits.R
\name{reviseSpeciesTraits}
\alias{reviseSpeciesTraits}
\title{Bootstraps rasters for testing significance on comparable rasters of different species or scenarios}
\usage{
reviseSpeciesTraits(
  speciesTable = NULL,
  updatedTable = NULL,
  destinationPath = tempdir(),
  returnTable = FALSE,
  species = c("BETU.PAP", "LARI.LAR", "PICE.GLA", "PICE.MAR", "PINU.BAN", "POPU.TRE",
    "PINU.CON")
)
}
\arguments{
\item{speciesTable}{\code{data.table}. Table to be updated.
If \code{returnTable == TRUE} this can be NULL.
In this case, the function will use the ORIGINAL species table traits (from LANDIS-II).}

\item{updatedTable}{\code{data.table}. Table used to update \code{speciesTable}.
If NULL, the function uses a default.}

\item{destinationPath}{character. Where to save the table. Default is \code{tempdir().}}

\item{returnTable}{logical. If TRUE, it uses the ORIGINAL species table traits from LANDIS-II. If FALSE, you have to supply your own table.}

\item{species}{character.  Species to update the table for.}
}
\value{
Table of updated tree species traits to be used in LandR biomass simulation
}
\description{
Bootstraps rasters for testing significance on comparable rasters of different species or scenarios
}
\author{
Tati Micheletti
}
