% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineStudyArea.R
\name{defineStudyArea}
\alias{defineStudyArea}
\title{Preparing study area for Canadian projects}
\usage{
defineStudyArea(
  testArea = NULL,
  specificTestArea = NULL,
  mapSubset = NULL,
  destinationFolder = tempdir()
)
}
\arguments{
\item{testArea}{Logical. Indicates if the test area should
be anything other than \code{NULL}. Default is \code{NULL}.}

\item{specificTestArea}{A character string with a province or territory name, or \code{'boreal'}.
(following Brandt \emph{et al.}, 2013). Default is \code{NULL}.}

\item{mapSubset}{If \code{specificTestArea} is supplied as \code{'boreal'}, this can be set
as a character string with a province or territory that is contained
in the boreal or \code{'Canada'} if the whole Canadian boreal is wanted.
Default is \code{NULL}.}

\item{destinationFolder}{Path to where to save downloaded files. Default is \code{tempdir()}.}
}
\description{
Downloads, reprojects, crops and masks to specific areas in Canada such as:
boreal, random areas, provinces and territories, or any of the last in the boreal.
}
\examples{
## rp is the whole North American boreal region
rp <- defineStudyArea(testArea = TRUE, specificTestArea = "boreal", mapSubset = NULL)

## Alberta
rp <- defineStudyArea(testArea = TRUE, specificTestArea = "Alberta", mapSubset = NULL)

\dontrun{
## Alberta inside boreal extension
rp <- defineStudyArea(testArea = TRUE, specificTestArea = "boreal", mapSubset = "Alberta")
}
}
\author{
Tati Micheletti
}
