% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineStudyArea.R
\name{defineStudyArea}
\alias{defineStudyArea}
\title{Preparing study area for Canadian projects}
\usage{
defineStudyArea(
  testArea = NULL,
  specificTestArea = NULL,
  mapSubset = NULL,
  destinationFolder = tempdir()
)
}
\arguments{
\item{testArea}{Logical. Indicates if the test area should
be anything other than `NULL`. Default is `NULL`.}

\item{specificTestArea}{A character string with a province or territory, or 'boreal'.
if boreal region (following Brandt et al., 2013) is wanted.
Default is `NULL`.}

\item{mapSubset}{If specificTestArea is supplied as 'boreal', this can be set
as a character string with a province or territory that is contained
in the boreal or 'Canada' if the whole Canadian boreal is wanted.
Default is `NULL`.}

\item{destinationFolder}{Path to where to save downloaded files. Default is `tempdir()`.}
}
\description{
Downloads, reprojects, crops and masks to speficic areas in canada such as:
boreal, random areas, provinces and territories, or any of the last in the
boreal.
}
\examples{
\dontrun{
## rp is the whole North American boreal region
rp <- defineStudyArea(testArea = TRUE, specificTestArea = "boreal", mapSubset = NULL)
}
\dontrun{
## Alberta
rp <- defineStudyArea(testArea = TRUE, specificTestArea = "Alberta", mapSubset = NULL)
}
\dontrun{
## Alberta inside boreal extension
rp <- defineStudyArea(testArea = TRUE, specificTestArea = "boreal", mapSubset = "Alberta")
}
}
\author{
Tati Micheletti
}
