% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLeadingVegetationType.R
\name{plotLeadingVegetationType}
\alias{plotLeadingVegetationType}
\title{Plots Leading Vegetation Type using cohortData and pixelGroupMap}
\usage{
plotLeadingVegetationType(
  dataPath,
  typeSim,
  colNA = "grey85",
  saveRAS = TRUE,
  overwrite = FALSE,
  sppColorVect,
  sppEquivCol,
  sppEquivalencies_CA
)
}
\arguments{
\item{dataPath}{character. Path to data}

\item{typeSim}{character. Which simulation is it? i.e. 'LandR_SCFM' | 'LandR.CS_fS'}

\item{colNA}{character. The color to use for NA.}

\item{saveRAS}{logical. Save the raster for posterior use?}

\item{overwrite}{logical.}

\item{sppColorVect}{named character vector with the species colors in hex format}

\item{sppEquivCol}{character with the column from sppEquivalencies_CA to use for the map}

\item{sppEquivalencies_CA}{sppEquivalencies_CA object from LandR (LandR::sppEquivalencies_CA) changed or not}
}
\value{
plot
}
\description{
Plots Leading Vegetation Type using cohortData and pixelGroupMap
}
\author{
Tati Micheletti
}
